experiments/mae_vit_high/requirements.txt
torch
torchvision
PyYAML
numpy
